Moves prolongate to lf::uscalfe
