+++ lib/lf/uscalfe/test/CMakeLists.txt
@@ -13,7 +13,7 @@ set(src
   lagr_fe_tests.cc
   mesh_function_fe_tests.cc
   mesh_function_grad_fe_tests.cc
-  
+  prolongation_tests.cc
 )
 
 add_executable(lf.uscalfe.test ${src})
